Make FFI code internal
