+++ internal/gpu/bridge.go
